command line flag handling using go-flags
While reviewing the PR to add version output to `httpstat`, I thought it
would be nice to have parity with `curl`'s flag. In addition to that,
with us aiming for a `1.0` release, it'd be nice to have assertions
around the command line argument parsing.

This changeset changes the command line flag parsing from the standard
library to that of `go-flags`. The way `go-flags` is consumed here is a
bit complicated, but it's so that we can abstract functionality away and
properly unit test all command line flag parsing.

This also fixes a bug in the error handling around the output file
creation. Previously, if the file couldn't be created it may have used
the wrong filename in the error message.